chore: refactor api/primitives.ts to display error toast on every failure
- refactor `request` method to use `handlePromise` instead of try/catch
- refactored `~/lib/utils` to a folder with barreled exports (for backwards compatibility)

files of note:
- `~/lib/utils/index.ts`
- `~/lib/service/api/primitives.ts`
