+++ ui/admin/app/lib/utils/assetUrl.ts
@@ -1,0 +1,1 @@
+export const assetUrl = (path: string) => `${import.meta.env.BASE_URL}${path}`;
