+++ ui/admin/app/lib/utils/getAliasFrom.ts
@@ -1,0 +1,5 @@
+export const getAliasFrom = (text: Nullish<string>) => {
+	if (!text) return "";
+
+	return text.toLowerCase().replace(/[^a-z0-9-]+/g, "-");
+};
