+++ ui/admin/app/lib/utils/pluralize.ts
@@ -1,0 +1,2 @@
+export const pluralize = (count: number, singular: string, plural: string) =>
+	`${count === 1 ? singular : plural}`;
